feat : primary-mypage navigation error 수정
