+++ src/screens/primary/primary-my-page/module/primary-mypage-content/index.ts
@@ -1,1 +1,1 @@
-export * from './mypage-content.module';
+export * from './primary-mypage-content.module';
