+++ src/screens/primary/primary-my-page/module/primary-mypage-header/index.ts
@@ -1,1 +1,1 @@
-export * from './mypage-header.module';
+export * from './primary-mypage-header.module';
