+++ src/screens/primary/primary-my-page/primary-my-page.screen.tsx
@@ -1,19 +1,30 @@
 import { Stack } from '@mobily/stacks';
-import { useNavigation } from '@react-navigation/native';
+import { BottomTabNavigationProp } from '@react-navigation/bottom-tabs';
+import { CompositeNavigationProp, RouteProp } from '@react-navigation/native';
 
 import {
-  PrimaryMyPageContentModule,
-  PrimaryMyPageScreenNavigatorProp,
-} from './module/primary-mypage-content';
+  PrimaryNavigatorParamLists,
+  PrimaryNavigatorProps,
+} from '../primary.navigator';
+
+import { PrimaryMyPageContentModule } from './module/primary-mypage-content';
 import { PrimaryMyPageHeaderModule } from './module/primary-mypage-header';
 
 import { palette } from '@/utils';
 
 type PrimaryMyPageScreenProps = {};
 
-export const PrimaryMyPageScreen = ({}: PrimaryMyPageScreenProps) => {
-  const navigation = useNavigation<PrimaryMyPageScreenNavigatorProp>();
+export type PrimaryMyPageScreenNavigatorProp = CompositeNavigationProp<
+  PrimaryNavigatorProps,
+  BottomTabNavigationProp<PrimaryNavigatorParamLists, 'PrimaryMyPageScreen'>
+>;
 
+export type PrimaryMyPageScreenRouteProp = RouteProp<
+  PrimaryNavigatorParamLists,
+  'PrimaryMyPageScreen'
+>;
+
+export const PrimaryMyPageScreen = ({}: PrimaryMyPageScreenProps) => {
   return (
     <Stack
       paddingX={24}
@@ -21,7 +32,7 @@ export const PrimaryMyPageScreen = ({}: PrimaryMyPageScreenProps) => {
       style={{ backgroundColor: palette['gray-100'], flex: 1 }}
       space={48}>
       <PrimaryMyPageHeaderModule />
-      <PrimaryMyPageContentModule navigation={navigation} />
+      <PrimaryMyPageContentModule />
     </Stack>
   );
 };
